+++ R/objects.R
@@ -4477,6 +4477,15 @@ ReadH5AD.H5File <- function(
   ...
 ) {
   CheckDots(...)
+  .Deprecated(
+    package = 'SeuratDisk',
+    msg = paste(
+      "Functionality for reading and writing H5AD files is being moved to SeuratDisk",
+      "For more details, please see https://github.com/mojaveazure/seurat-disk",
+      "and https://mojaveazure.github.io/seurat-disk/index.html",
+      sep = "\n"
+    )
+  )
   # Pull assay data
   # If X is an H5D, assume scaled
   # Otherwise, if file$exists(name = 'raw'), assume X is normalized
@@ -6001,6 +6010,15 @@ WriteH5AD.Seurat <- function(
   overwrite = FALSE,
   ...
 ) {
+  .Deprecated(
+    package = 'SeuratDisk',
+    msg = paste(
+      "Functionality for reading and writing H5AD files is being moved to SeuratDisk",
+      "For more details, please see https://github.com/mojaveazure/seurat-disk",
+      "and https://mojaveazure.github.io/seurat-disk/index.html",
+      sep = "\n"
+    )
+  )
   message("WriteH5AD is not currently operational, please use as.loom")
   .NotYetImplemented()
   if (!PackageCheck('hdf5r', error = FALSE)) {